move runepouch rune enum to client/game
