+++ runelite-client/src/main/java/net/runelite/client/game/RunepouchRune.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.runepouch;
+package net.runelite.client.game;
 
 
 import com.google.common.collect.ImmutableMap;
@@ -52,7 +52,7 @@ import static net.runelite.api.ItemID.STEAM_RUNE;
 import static net.runelite.api.ItemID.WATER_RUNE;
 import static net.runelite.api.ItemID.WRATH_RUNE;
 
-public enum Runes
+public enum RunepouchRune
 {
 	AIR(1, AIR_RUNE),
 	WATER(2, WATER_RUNE),
@@ -85,25 +85,25 @@ public enum Runes
 	@Setter
 	private BufferedImage image;
 
-	private static final Map<Integer, Runes> runes;
+	private static final Map<Integer, RunepouchRune> runes;
 
 	static
 	{
-		ImmutableMap.Builder<Integer, Runes> builder = new ImmutableMap.Builder<>();
-		for (Runes rune : values())
+		ImmutableMap.Builder<Integer, RunepouchRune> builder = new ImmutableMap.Builder<>();
+		for (RunepouchRune rune : values())
 		{
 			builder.put(rune.getId(), rune);
 		}
 		runes = builder.build();
 	}
 
-	Runes(int id, int itemId)
+	RunepouchRune(int id, int itemId)
 	{
 		this.id = id;
 		this.itemId = itemId;
 	}
 
-	public static Runes getRune(int varbit)
+	public static RunepouchRune getRune(int varbit)
 	{
 		return runes.get(varbit);
 	}
